Change matcher to use templates
The user sets the template to use with `--match`.
The template must evaluate to a boolean value which is parsable by
`strconv.ParseBool`.
After the template executes we read the value from the buffer to
determine if the the resource should be included or not.

This includes 2 extra template functions that users can use:

- "regexp" - evalutes the regex against the provided string. e.g.
  `{{regexp .Path "bin/(sh|bash)$"}}`.
- "exec" - executes the provided command+args, exit code 0 means match,
  exit code 1 means not a match, exit code greater than 1 means error.
  e.g. `{{exec "/bin/sh" "-c" "[ \"$(jq .path <@0)\" eq \"/bin/sh\" ]"}}`

The match template is executed for each resource path found.